pyproject.toml
[tool.poetry]
name = "langchain-exploration"
version = "0.1.0"
description = "Exploring lang chain and lang graph"
authors = ["hobbyist-programmer-ar <hobbyistprogrammer@outlook.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
langchain-openai = "^0.2.0"
python-dotenv = "^1.0.1"
langchain = "^0.3.27"
langchain-community = "^0.3.27"
langchain-anthropic = "^0.3.18"
langchain-google-genai = "^2.1.8"
langchain-google-firestore = "^0.4.0"
firestore = "^0.0.8"
chromadb = "^0.5.0"
tiktoken = "^0.7.0"
sentence-transformers = "^3.0.0"
bs4 = "^0.0.2"
firecrawl-py = "^0.0.13"
langchainhub = "^0.1.21"
wikipedia = "^1.4.0"
tavily-python = "^0.3.3"
langchain-ollama = "^0.3.6"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
